Fix compile warnings
We are now more pedantic on compile warnings and
fix them, at least in our code.
Pluginlib spams a little, but that's fine for now.

+++ joint_to_cartesian_controller/include/joint_to_cartesian_controller/JointControllerAdapter.h
@@ -75,7 +75,7 @@ class JointControllerAdapter : public hardware_interface::RobotHW
 
   private:
     //! Number of actuated joints
-    int m_number_joints;
+    size_t m_number_joints;
 
     //! Actuated joints in order from base to tip
     std::vector<std::string> m_joint_names;
